+++ packages/network-controller/tests/network-client/rpc-failover.ts
@@ -1,10 +1,10 @@
 import { ConstantBackoff } from '@metamask/controller-utils';
 import type { Hex } from '@metamask/utils';
 
-import { ignoreRejection } from '../../../../tests/helpers';
-import { buildRootMessenger } from '../helpers';
-import type { MockRequest, MockResponse, ProviderType } from './helpers';
-import { withMockedCommunications, withNetworkClient } from './helpers';
+import { ignoreRejection } from '../../../../tests/helpers.js';
+import { buildRootMessenger } from '../helpers.js';
+import type { MockRequest, MockResponse, ProviderType } from './helpers.js';
+import { withMockedCommunications, withNetworkClient } from './helpers.js';
 
 /**
  * Tests for RPC failover behavior.